FITS operations moved to separate file
